Rename MenuSshLogin to MenuLogin
