+++ packages/shared/components/MenuLogin/index.ts
@@ -14,14 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export type LoginItem = {
-  url: string;
-  login: string;
-};
-
-export type MenuSshLoginProps = {
-  onOpen: () => LoginItem[];
-  onSelect: (e: React.SyntheticEvent, login: string) => void;
-  anchorOrigin?: any;
-  transformOrigin?: any;
-};
+import MenuLogin from './MenuLogin';
+export default MenuLogin;
+export * from './types';
